requirements.txt
PyQt5>=5.15.0
ultralytics>=8.0.0
PyYAML>=6.0
numpy>=1.20.0
opencv-python>=4.5.0

